+++ lib/input/gestures/move_camera_using_gestures.dart
@@ -63,5 +63,5 @@ mixin MoveCameraUsingGesture on GameComponent {
   }
 
   @override
-  bool get isVisible => true;
+  bool get isVisibleInScreen => true;
 }