+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - adds `spriteOffeset` in `UseSprite`.
 - adds `spriteAnimationOffset` in `UseSpriteAnimation`.
 - adds `spriteAnimationOffset` in `SimpleDirectionAnimation`.
+- `BlockMovementCollision` improvements. BREAKING CHANGE: Method `onBlockedMovement` now return not null `Direction`.
+- BREAKING CHANGE: the param `isVisible` was renamed to `isVisibleInScreen`.
+- Adds param `visible` in `GameComponent`.
+
 
 ## 3.0.6
 - Adds bool `paused` in `BonfireGameInterface`.